separate out getBuiltInIntegrationConfigs
